+++ lib/credentials.dart
@@ -1,4 +1,5 @@
 export 'src/credentials/credential_operations.dart';
+export 'src/credentials/credential_signer.dart';
 export 'src/credentials/presentation_exchange.dart';
 export 'src/credentials/revocation.dart';
 export 'src/credentials/verifiable_credential.dart';